+++ changelog.md
@@ -20,8 +20,11 @@
 - [#3559](https://github.com/ignite/cli/pull/3559) Bump network plugin version to `v0.1.1`
 - [#3522](https://github.com/ignite/cli/pull/3522) Remove indentation from `chain serve` output
 - [#3601](https://github.com/ignite/cli/pull/3601) Update ts-relayer version to `0.10.0`
+- [#3658](https://github.com/ignite/cli/pull/3658) Rename Marshaler to Codec in EncodingConfig
 - [#3653](https://github.com/ignite/cli/pull/3653) Add "app" extension to plugin binaries
 - [#3656](https://github.com/ignite/cli/pull/3656) Disable Go toolchain download
+- [#3662](https://github.com/ignite/cli/pull/3662) Refactor CLI "plugin" command to "app"
+- [#3669](https://github.com/ignite/cli/pull/3669) Rename `plugins` config file to `igniteapps`
 
 ### Fixes
 
@@ -30,6 +33,7 @@
 - [#3610](https://github.com/ignite/cli/pull/3610) Fix overflow issue of cosmos faucet in `pkg/cosmosfaucet/transfer.go` and `pkg/cosmosfaucet/cosmosfaucet.go`
 - [#3618](https://github.com/ignite/cli/pull/3618) Fix TS client generation import path issue
 - [#3631](https://github.com/ignite/cli/pull/3631) Fix unnecessary vue import in hooks/composables template
+- [#3661](https://github.com/ignite/cli/pull/3661) Change `pkg/cosmosanalysis` to find Cosmos SDK runtime app registered modules
 
 ## [`v0.27.0`](https://github.com/ignite/cli/releases/tag/v0.27.0)